Step-3: Implement a method to do the actual work
In this step we implement the method that actually does the work for our
plugin... in this case, it counts nodes.  The code iterates through the
list of nodes passed in, gets their count from the document, and adds it
to the beacon.

We use the BOOMR.addVar method to add a variable to the beacon.

We've also added a new configurable parameter called `prefix`.  It's a
good idea for all plugins to prefix the variables they add to the beacon
with a namespace.  This helps avoid conflicts across plugins.

Unfortunately early plugins did not follow this convention, so we've got
a few variables that are badly named and need to stay that way to be
backwards compatible.