Step-4: Tell boomerang to send the beacon
Now that we've done what we needed to do, we tell boomerang that we're
ready to send the beacon.

Boomerang may not send the beacon out immediately.  It will first check
the is_complete() method of all plugins (this one included), and if they
all return true, the beacon will be sent.  If any one of them returns
false, the beacon will be deferred until sendBeacon is called again.

Thus, it is important that your plugin always call sendBeacon regardless
of whether it executed successfully or not.  It is a good practice to
set a failure state using addVar if appropriate.

We also need to set complete to true before calling sendBeacon so that
our own is_complete method returns true when called.

Forgetting to do one or both of these has resulted in bugs in the past.