Step-5: Execute our method after onload
We can execute our method at any time, but it's probably best to wait
until the page_ready event has fired so we can be sure we're measuring
nodes from the complete page.

In this case we could have also just waited for the dom_loaded event.

Boomerang fires the following events:
- page_ready: by default this is fired when the page's onload event
  fires, but the page developer can override this
- page_unload: this fires twice.  Once when the onbeforeunload event
  fires on the window, and the second time when either the pagehide or
  the unload events fire (but not on both).
- dom_loaded: fired at the same time as DOMContentLoaded on browsers
  that support it
- click: fired when a click event happens on the page
- visibility_changed: fires when the page's visibility state changes
- before_beacon: fires just before the beacon is about to be sent

See the boomerang documentation for full instructions