Step-6: Use the actual host object
There are 3 ways in which boomerang may be loaded onto a page.
1. A direct static script node
2. A dynamic script node
3. A dynamic script node inside a dynamic iframe

In the first two cases, the window object points to the host window,
however, in case 3, the window object is the iframe in which boomerang
is loaded.  This is not the same as the host window.

To work around this, we use BOOMR.window rather than window whenever we
need to refer to the host window object.  Note however that BOOMR.window
is only guaranteed to be set after init has been called, so it cannot be
used in a global execution scope.

See http://www.lognormal.com/blog/2012/12/12/the-script-loader-pattern/
for how and why we use method #3.